modify f

+++ printf/ft_printf_bonus/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: ylee <ylee@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/03 16:21:31 by ylee              #+#    #+#             */
-/*   Updated: 2021/01/13 15:26:34 by ylee             ###   ########.fr       */
+/*   Updated: 2021/01/18 13:03:26 by ylee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,6 +69,6 @@ char			*print_type_n(va_list ap, t_arg *arg);
 char			*print_type_f(va_list ap, t_arg *arg);
 char			*print_type_g(va_list ap, t_arg *arg);
 char			*print_type_e(va_list ap, t_arg *arg);
-char			*testf(float num);
+char			*testf(double num);
 
 #endif
