+++ printf/ft_printf_bonus/print_type_f.c
@@ -6,7 +6,7 @@
 /*   By: ylee <ylee@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/02 09:29:57 by ylee              #+#    #+#             */
-/*   Updated: 2021/01/15 15:30:51 by ylee             ###   ########.fr       */
+/*   Updated: 2021/01/18 13:02:15 by ylee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 char	*print_type_f(va_list ap, t_arg *arg)
 {
 	char	*result;
-	float	num;
+	double	num;
 
 	result = NULL;
 	num = 0;
 	if (arg->specifier == 'f')
-		num = (float)va_arg(ap, double);
+		num = va_arg(ap, double);
 	result = testf(num);
 	arg->tmp_len = ft_strlen(result);
 	return (result);
